requirements.txt
python-telegram-bot==22.0
SQLAlchemy==2.0.36
python-dotenv==1.0.1